Move certificate generator into LegAid
