+++ LegAid/requirements.txt
@@ -1,4 +1,9 @@
 streamlit
-pandas
 openai
+pandas
 pdfminer.six
+python-docx
+python-dateutil
+openpyxl
+pillow
+pytesseract
